backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
doc2text==0.2.4
docx2python==2.10.1
docx2txt==0.8
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
joblib==1.3.2
libmagic==1.0
lxml==5.2.1
mime==0.1.0
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
openai==0.28.0
packaging==24.0
pandas==2.2.1
paragraphs==0.2.1
pillow==10.3.0
py4j==0.10.9.5
pydantic==2.6.3
pydantic_core==2.16.3
PyPDF2==3.0.1
pyspark==3.3.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-magic==0.4.3
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
